Define the calendar service methods
The calendar service uses the same url, and an additional parameter each
time for the year, week and month activites. The responses, from each
URL are different. Well, year is different from the others. For that
reason, I have created three methods. It seem's more obvious, and
simpler.

I have not opted for a date input, just an integer for each parameter.
Accepting a date would be possible.